debug

+++ MuonAnalysis/python/DATA.py
@@ -48,7 +48,7 @@ def prepareSet( data, label_dict):
 	y_test = target.to_numpy()
 	
 
-	y_test = np.array([np.array(i) for i in y_test])
+	y_test = np.array([np.array(i).astype(int) for i in y_test])
 	pt_test = pt_test.to_numpy()
 	return x_test,y_test,pt_test
